Break PD hidden state import cycle
